feat: Coupon 추가 로직 작성

+++ src/main/java/org/c4marathon/assignment/domain/discountpolicy/repository/DiscountPolicyRepository.java
@@ -3,7 +3,7 @@ package org.c4marathon.assignment.domain.discountpolicy.repository;
 import org.c4marathon.assignment.domain.discountpolicy.entity.DiscountPolicy;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DiscountPolicyRepository extends JpaRepository<DiscountPolicy, Integer> {
+public interface DiscountPolicyRepository extends JpaRepository<DiscountPolicy, Long> {
 
 	boolean existsByName(String name);
 }
